+++ src/tavern/mod.rs
@@ -3,4 +3,6 @@
 #[cfg(feature = "openssl")]
 pub mod openssl;
 pub mod option;
+#[cfg(feature = "rustls")]
+pub mod rustls;
 pub mod tcp_udp;